+++ src/e2e/codegen/typescript/mod.rs
@@ -100,9 +100,7 @@ impl E2eCodegen for TypeScriptCodegen {
                 &f.tags,
                 &f.input,
             );
-            cc.args
-                .iter()
-                .any(|a| a.arg_type == "file_path" || a.arg_type == "bytes")
+            super::file_inputs::fixture_uses_test_documents(f, cc, type_defs)
         });
 
         files.push(GeneratedFile {